LayoutTests/fast/shadow-dom/scoped-custom-element-registry-disabled.html

<!DOCTYPE html> <!-- webkit-test-runner [ ScopedCustomElementRegistryEnabled=false ] -->
<script src="../../resources/js-test.js"></script>
<script>
shouldBeTrue("document.createElement('div').attachShadow({mode: 'open'}) instanceof ShadowRoot");
</script>
