LayoutTests/fast/shadow-dom/scoped-custom-element-registry-disabled-expected.txt

PASS document.createElement('div').attachShadow({mode: 'open'}) instanceof ShadowRoot is true
PASS successfullyParsed is true

TEST COMPLETE
